# docs/swagger.yaml
basePath: /
definitions:
...
  models.Movie:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Seat:
    properties:
      id:
        type: string
      seat_number:
        type: string
      showtime_id:
        type: string
      status:
        $ref: '#/definitions/models.SeatStatus'
    type: object
  models.SeatStatus:
    enum:
    - AVAILABLE
    - LOCKED
    - BOOKED
    type: string
    x-enum-varnames:
    - SeatAvailable
    - SeatLocked
    - SeatBooked
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a movie ticket booking system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Movie Ticket Booking API
  version: "1.0"
paths:
...
  /api/movies:
    get:
      description: Fetch a list of all movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all movies
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Add a new movie to the system (Admin only)
      parameters:
      - description: Movie Data
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new movie
      tags:
      - Movies
  /api/movies/{id}:
    delete:
      description: Delete a movie from the system by ID (Admin only)
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a movie
      tags:
      - Movies
    get:
      description: Fetch a single movie by its ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a movie by ID
      tags:
      - Movies
    put:
      consumes:
      - application/json
      description: Update movie details by ID (Admin only)
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a movie
      tags:
      - Movies
...
  /api/seats:
    get:
      description: Fetch a list of all seats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all seats
      tags:
      - Seats
    post:
      consumes:
      - application/json
      description: Add a new seat to the system (Admin only)
      parameters:
      - description: Seat Data
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/models.Seat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new seat
      tags:
      - Seats
  /api/seats/{id}:
    delete:
      description: Delete a seat from the system by ID (Admin only)
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a seat
      tags:
      - Seats
    get:
      description: Fetch a single seat by its ID
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seat'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a seat by ID
      tags:
      - Seats
    put:
      consumes:
      - application/json
      description: Update seat details by ID (Admin only)
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a seat
      tags:
      - Seats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"